WIP zeropad all the sections
Without this, during fromelf convertion to bin, _data section
is merged to the previous one. It is not aligned, as well as placed
by the wrong address in the resulted bin (and runtime).
Zeropad all the sections to be sure they are placed in the right places.
This increases the binary to 1MB+ but it is starts up to the functional
Dom0 console.

We stil have a lot of linkage warnings, and not eliminated GCC objcopy and nm.